Add files via upload
Welcome to the Python (NumPy and Pandas) folder! This is a comprehensive collection of important topics in Python programming language with a specific focus on the NumPy and Pandas libraries.

The topics covered in this folder include but are not limited to:

Python basics (variables, data types, loops, conditionals, functions, etc.)
NumPy arrays (creating arrays, indexing and slicing, array operations, etc.)
Pandas dataframes (creating dataframes, indexing and selecting data, data manipulation, etc.)
Data cleaning and preprocessing with NumPy and Pandas
Data analysis with NumPy and Pandas
Data visualization with Matplotlib and Seaborn libraries
Time series analysis with Pandas
Each topic is presented in a clear and concise manner, with examples and exercises to reinforce your understanding. This folder also includes resources such as cheat sheets, formula sheets, and Python software tutorials to assist you in your programming and data analysis.

The goal of this folder is to provide you with a comprehensive understanding of Python programming language, specifically with the NumPy and Pandas libraries, that can be applied to various fields such as data science, machine learning, finance, and more. We hope you find this folder useful and welcome any feedback or suggestions for improvement.